+++ shared-android/src/main/java/dev/olog/shared/android/theme/HasImageShape.kt
@@ -8,6 +8,6 @@ interface HasImageShape {
 }
 
 enum class ImageShape {
-    RECTANGLE, ROUND
+    RECTANGLE, ROUND, CUT_CORNER
 }