+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - added volume controls to player fragment
 - made cover colored in dark mode
 - added is_podcast override in edit track
+- added cut corner image shape
+- added image stylizer
 
 ## 2.1.1
 - fixed all tracks sort order by date